Python scripts from the notebooks
Since I need the Python files in order to use the modules elsewhere,
I have created a script that converts the notebooks into Python and
then move those files into the src folder; this should be enough for
my needs at the moment. I might have to create some tests to ensure
the consistency of the generated scripts, but I'll do this later.