Minor fixes to the Petri Net graph
There were two issues with the graphs being created: the first is
that I was not handling the offset for the transitions, so they
were using th same range as places; the second was that the second
set of edges should be flipped, since we are going from transitions
to places. These two problems have been fixed and I should be able
to create some graphs now.