Implementing basic DGLDataset
Since I am having some problems with PyGeometric, having a branch
with DGL is a good idea. This basic implementation is already
working in the sense that it reads the data correctly, but I have
not tested the learning process with it, nor have I implemented the
padding of features for the dataset.