PyGeometric Example
This is an example file source from the PyGeometric project that does exactly
what I want: it uses a GCN model with node and edge features and outputs a
multi-class choice; the only difference to mine is that instead of multiple
classes I will have a single one, which means a regression. I should be able to
test my dataset with multiple configurations using this.