Simplification of dataset creation
Instead of padding the node features so that all have the same
size, I can just reduce the number to a scalar. I have alredy done
some simple tests and checked that, indeed, the data loader works;
the issue now is in loading the correct number of nodes. The
solutions are either implementing a Dataset so that I can use a
node sampler through PyG or pad the graphs with empty nodes.