fixes and tests
I tested both of these models and they are running without issues; the only
things missing are just some basic handling of the output, but this is a
quality-of-life improvement. The next step will be adding the MLP and CNN
models, which are both simpler and harder to do: simpler in that they are
homogeneous in the type of layer used, and harder because I have to do an
additional data conversion before input.