train/test loop
The basic train/test/next loop is done, now I just need to narrow
the number of combinations. I am thinking about just testing up
to 5 GCN layers and then 4 MLP readout layers for this; I might
also use `optuna` for this, since I learned that using compilation
with it is a mistake.