Off-by-one errors and other minor issues
There were some small problems on `datasets.py`, such as me
assuming that the PyTorch sum was identical to that of Numpy and
the careless consumption of iterators when I should materialize
it into a list; there was also a minor off by one error on
`data_parser.py`. These have all been fixed.